inmar/requirements.txt
amqp==1.4.9
anyjson==0.3.3
backports-abc==0.5
billiard==3.3.0.23
boto==2.48.0
celery==3.1.25
certifi==2017.7.27.1
Django==1.11.5
django-celery==3.2.1
django-redis==4.8.0
djangorestframework==3.6.4
drf-nested-routers==0.90.0
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
kombu==3.0.37
psycopg2==2.7.3.1
pytz==2017.2
redis==2.10.6
singledispatch==3.4.0.3
six==1.10.0
sockjs-tornado==1.0.3
tornado==4.5.2
django-oauth-toolkit==0.10.0
